refactor(hooks): share hook metadata parsing helpers
