+++ src/commands/hook-events/hook-utils.ts
@@ -1,0 +1,35 @@
+export function normalizeWhitespace(value: string): string {
+	return value.replace(/\s+/g, " ").trim();
+}
+
+export function asRecord(value: unknown): Record<string, unknown> | null {
+	if (!value || typeof value !== "object" || Array.isArray(value)) {
+		return null;
+	}
+	return value as Record<string, unknown>;
+}
+
+export function readStringField(record: Record<string, unknown>, key: string): string | null {
+	const value = record[key];
+	if (typeof value !== "string") {
+		return null;
+	}
+	const normalized = normalizeWhitespace(value);
+	return normalized.length > 0 ? normalized : null;
+}
+
+export function readNestedString(record: Record<string, unknown>, path: string[]): string | null {
+	let current: unknown = record;
+	for (const key of path) {
+		const candidate = asRecord(current);
+		if (!candidate || !(key in candidate)) {
+			return null;
+		}
+		current = candidate[key];
+	}
+	if (typeof current !== "string") {
+		return null;
+	}
+	const normalized = normalizeWhitespace(current);
+	return normalized.length > 0 ? normalized : null;
+}
