+++ src/commands/hook-events/kiro-hook-events.ts
@@ -1,46 +1,11 @@
 import type { RuntimeHookEvent, RuntimeTaskHookActivity } from "../../core/api-contract";
+import { asRecord, normalizeWhitespace, readNestedString, readStringField } from "./hook-utils";
 
 interface NormalizeKiroHookMetadataArgs {
 	event: RuntimeHookEvent;
 	payload: Record<string, unknown> | null;
 	flagMetadata: Partial<RuntimeTaskHookActivity>;
 	sourceHint?: string | null;
-}
-
-function normalizeWhitespace(value: string): string {
-	return value.replace(/\s+/g, " ").trim();
-}
-
-function asRecord(value: unknown): Record<string, unknown> | null {
-	if (!value || typeof value !== "object" || Array.isArray(value)) {
-		return null;
-	}
-	return value as Record<string, unknown>;
-}
-
-function readStringField(record: Record<string, unknown>, key: string): string | null {
-	const value = record[key];
-	if (typeof value !== "string") {
-		return null;
-	}
-	const normalized = normalizeWhitespace(value);
-	return normalized.length > 0 ? normalized : null;
-}
-
-function readNestedString(record: Record<string, unknown>, path: string[]): string | null {
-	let current: unknown = record;
-	for (const key of path) {
-		const candidate = asRecord(current);
-		if (!candidate || !(key in candidate)) {
-			return null;
-		}
-		current = candidate[key];
-	}
-	if (typeof current !== "string") {
-		return null;
-	}
-	const normalized = normalizeWhitespace(current);
-	return normalized.length > 0 ? normalized : null;
 }
 
 function extractKiroToolInput(payload: Record<string, unknown> | null): Record<string, unknown> | null {
